requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apt-xapian-index==0.47
asgiref==3.3.1
asn1crypto==0.24.0
beautifulsoup4==4.6.0
catfish==1.4.4
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.5
django-crispy-forms==1.10.0
django-filter==2.4.0
django-mysql==3.10.0
django-phone-field==1.8.1
djangorestframework==3.12.2
filelock==3.0.12
freeze-requirements==0.5.3
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==3.4.0
importlib-resources==5.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lightdm-gtk-greeter-settings==1.2.2
lxml==4.2.1
menulibre==2.2.0
mugshot==0.4.0
mysql-connector==2.2.9
mysqlclient==2.0.3
netifaces==0.10.4
olefile==0.45.1
PAM==0.4.2
pexpect==4.2.1
Pillow==5.1.0
powerline-status==2.6
psutil==5.4.2
psycopg2==2.8.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
PyMySQL==0.8.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2020.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sessioninstaller==0.0.0
sh==1.14.1
six==1.11.0
sqlparse==0.4.1
systemd-python==234
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
virtualenv==20.4.2
webencodings==0.5
xkit==0.0.0
zipp==3.4.0
